+++ packages/react/src/FlowgladProvider.tsx
@@ -11,6 +11,7 @@ import {
 import { validateUrl } from './utils'
 import { FlowgladThemeProvider } from './FlowgladTheme'
 import { FlowgladThemeConfig } from './lib/themes'
+import { CustomerBillingDetails } from '@flowglad/types'
 
 const queryClient = new QueryClient()
 
@@ -18,20 +19,41 @@ export interface Appearance {
   darkMode?: boolean
 }
 
-export const FlowgladProvider = ({
-  children,
-  loadBilling,
-  serverRoute,
-  requestConfig,
-  theme,
-}: {
+export interface LoadedFlowgladProviderProps {
   children: React.ReactNode
-  appearance?: Appearance
   requestConfig?: RequestConfig
   serverRoute?: string
   loadBilling: boolean
   theme?: FlowgladThemeConfig
-}) => {
+}
+
+interface DevModeFlowgladProviderProps {
+  __devMode: true
+  theme?: FlowgladThemeConfig
+  billingMocks: CustomerBillingDetails
+  children: React.ReactNode
+}
+
+export type FlowgladProviderProps =
+  | LoadedFlowgladProviderProps
+  | DevModeFlowgladProviderProps
+
+export const FlowgladProvider = (props: FlowgladProviderProps) => {
+  if ('__devMode' in props) {
+    return (
+      <QueryClientProvider client={queryClient}>
+        <FlowgladContextProvider
+          __devMode
+          billingMocks={props.billingMocks}
+        >
+          {props.children}
+        </FlowgladContextProvider>
+      </QueryClientProvider>
+    )
+  }
+
+  const { serverRoute, loadBilling, requestConfig, theme, children } =
+    props as LoadedFlowgladProviderProps
   validateUrl(serverRoute, 'serverRoute', true)
   return (
     <QueryClientProvider client={queryClient}>