+++ packages/server/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @flowglad/server
 
+## 0.10.18
+
+### Patch Changes
+
+- add devmode support to FlowgladContext
+- Updated dependencies
+  - @flowglad/shared@0.10.18
+
 ## 0.10.17
 
 ### Patch Changes